SIG-873 split routers in v0 and v1
