+++ api/app/signals/apps/signals/urls.py
@@ -1,16 +1,12 @@
 from django.urls import include, path
 
-from signals.apps.signals import routers
 from signals.apps.signals.v0 import views as v0_views
-from signals.apps.signals.v1.views import (
-    GeneratePdfView,
-    MainCategoryViewSet,
-    PrivateSignalViewSet,
-    SubCategoryViewSet
-)
+from signals.apps.signals.v0.routers import SignalsRouterVersion0
+from signals.apps.signals.v1 import views as v1_views
+from signals.apps.signals.v1.routers import SignalsRouterVersion1
 
 # API Version 0
-signal_router_v0 = routers.SignalsRouterVersion0()
+signal_router_v0 = SignalsRouterVersion0()
 signal_router_v0.register(r'signal/image', v0_views.SignalImageUpdateView, base_name='signal-img')
 signal_router_v0.register(r'signal', v0_views.SignalViewSet, base_name='signal')
 signal_router_v0.register(r'auth/signal', v0_views.SignalAuthViewSet, base_name='signal-auth')
@@ -21,24 +17,24 @@ signal_router_v0.register(r'auth/priority', v0_views.PriorityAuthViewSet, base_n
 signal_router_v0.register(r'auth/note', v0_views.NoteAuthViewSet, base_name='note-auth')
 
 # API Version 1
-signal_router_v1 = routers.SignalsRouterVersion1()
+signal_router_v1 = SignalsRouterVersion1()
 signal_router_v1.register(r'public/terms/categories',
-                          MainCategoryViewSet,
+                          v1_views.MainCategoryViewSet,
                           base_name='category')
 signal_router_v1.register(r'private/signals',
-                          PrivateSignalViewSet,
+                          v1_views.PrivateSignalViewSet,
                           base_name='private-signals')
 
 # Appending extra url route for sub category detail endpoint.
 signal_router_v1.urls.append(
     path('public/terms/categories/<str:slug>/sub_categories/<str:sub_slug>',
-         SubCategoryViewSet.as_view({'get': 'retrieve'}),
+         v1_views.SubCategoryViewSet.as_view({'get': 'retrieve'}),
          name='sub-category-detail'),
 )
 
 signal_router_v1.urls.append(
     path('pdf/SIA-<int:signal_id>.pdf',
-         GeneratePdfView.as_view(),
+         v1_views.GeneratePdfView.as_view(),
          name='signal-pdf-download'),
 )
